experimental/requirements.txt
langchain==0.1.0
langchain-openai==0.0.8
openai==1.12.0
python-dotenv==1.0.0
yfinance==0.2.25
requests==2.31.0
pandas==2.1.4